+++ src/App.jsx
@@ -2,8 +2,8 @@
 import React from 'react'
 import { Routes, Route } from 'react-router-dom'
 /* Maneth */
-import Home from './pages/Home.jsx';
-import Login from './pages/Login.jsx';
+import Home from './pages/onlinestore/Home.jsx';
+import Login from './pages/onlinestore/Login.jsx';
