+++ src/components/footer/Footer.jsx
@@ -3,16 +3,16 @@ import React from "react";
 import FooterSocialLinks from "./FooterSocialLinks";
 import FooterLinks from "./FooterLinks";
 import FooterMap from "./FooterMap";
-import FooterContact from "./FooterContact";
+import FooterFeedback from "./FooterFeedback.jsx";
 
 const Footer = () => {
   return (
     <>
-      <div className="flex flex-row mt-[5%] w-full bg-footer min-h-[350px]">
+      <div className="flex flex-row mt-[5%] w-full bg-footer min-h-[350px] bg-bgc">
         <FooterSocialLinks />
         <FooterLinks />
         <FooterMap />
-        <FooterContact />
+        <FooterFeedback />
       </div>
       <center>
         <label>© 2022 Serendib Fashions Ltd.</label>
