+++ src/components/footer/FooterFeedback.jsx
@@ -1,0 +1,13 @@
+// eslint-disable-next-line no-unused-vars
+import React from "react";
+
+const FooterFeedback = () => {
+  return (
+    <div className="flex flex-col justify-evenly ml-[2%] w-[24%] mr-[2%]">
+        <h1>Like, dislike or having trouble?</h1>
+        
+    </div>
+  );
+};
+
+export default FooterFeedback;
