+++ src/components/footer/FooterMap.jsx
@@ -6,7 +6,7 @@ const FooterMap = () => {
     <div className="w-[26%]">
       <iframe
         className="mt-[5%] ml-[5%] w-[95%] h-[90%]"
-        src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d3960.994861527963!2d79.87481577576362!3d6.891216893107893!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x3ae25a103d225647%3A0xef1259856066f0bf!2sNational%20Blood%20Transfusion%20Service!5e0!3m2!1sen!2ssg!4v1705701712880!5m2!1sen!2ssg"
+        src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d126791.68343579641!2d79.87466096878049!3d6.740538712541805!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x3ae24dc585a53baf%3A0xde9b252e8b6625a4!2sSerendib%20Fashion%20(Pvt)%20ltd!5e0!3m2!1sen!2slk!4v1711857526903!5m2!1sen!2slk"
         loading="lazy"
         referrerPolicy="no-referrer-when-downgrade"
       ></iframe>
