+++ src/components/navbar/NavbarLogo.jsx
@@ -5,13 +5,13 @@ import { Link } from "react-router-dom";
 const Logo = () => {
   return (
     <div>
-      <Link to="/" className="flex flex-row w-[25rem]">
+      <Link to="/" className="flex flex-row mt-2 w-full">
           <img
             src="/Logo1.png"
             alt="logo"
-            className="w-[4rem] h-[4rem] ml-[1rem] mr-[2rem]"
+            className="w-full h-[3rem] ml-[1rem] mr-[2rem]"
           />
-          <img src="/Logo2.png" alt="logo" className="w-[19rem] " />
+          <img src="/Logo2.png" alt="logo" className="w-[13rem] hidden lg:block" />
       </Link>
     </div>
   );
