+++ src/pages/onlinestore/Home.jsx
@@ -1,9 +1,9 @@
 // eslint-disable-next-line no-unused-vars
 import React from "react";
-import HomeSlideShow from "../components/home/HomeSlideShow.jsx";
-import Footer from "../components/footer/Footer.jsx";
-import Navbar from "../components/navbar/CustomerNavbar.jsx";
-import HorizontalScroll from "../components/HorizontalScroll.jsx";
+import HomeSlideShow from "../../components/home/HomeSlideShow.jsx";
+import Footer from "../../components/footer/Footer.jsx";
+import Navbar from "../../components/navbar/CustomerNavbar.jsx";
+import HorizontalScroll from "../../components/HorizontalScroll.jsx";
 
 
 const Home = () => {
